Created a skeleton App and wired up React+Redux
The skeleton of the App comes with some basic tests for each component
involved in the React-Redux infrastructure.

Added a navigation Header Component to the App, this component provides
the Link to the Components the user will be able to navigate.
If the user is logged in, then 2 more components will show inside the
navigation, these are the Catalogue and the Profile page.

Catalogue and Profile Components, are enclosed into a HOC which
reads the current user status and either provides the passed Component
or kicks the user back to the home